+++ app/Console/Kernel.php
@@ -2,6 +2,7 @@
 
 namespace App\Console;
 
+use App\Console\Commands\SendAdminMail;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
 
@@ -13,7 +14,7 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        //
+        // SendAdminMail::class
     ];
 
     /**
@@ -23,7 +24,9 @@ class Kernel extends ConsoleKernel
      * @return void
      */
     protected function schedule(Schedule $schedule)
-    {
+    {   
+        // $schedule->command('send:tweetstats')->everyMinute();
+        $schedule->command('send:tweetstats')->dailyAt('21:00');
         $schedule->command('twitter:scan')->everyFifteenMinutes();
         $schedule->command('twitter:screen 100')->everyMinute();
         $schedule->command('twitter:duplicates 100')->everyMinute();