+++ CHANGELOG.md
@@ -16,3 +16,8 @@
 
 Made with love by [icrewsystems](https://icrewsystems.com)
 
+### Acitvity module 
+- Activity module added 
+
+### Refreal system 
+- work in progress